requirements.txt
python-telegram-bot==22.6
pyTelegramBotAPI
openai
requests
psutil
psycopg2-binary